requirements.txt
Pillow
numpy
netCDF4
